fix: separate portable and rendering tests
