+++ skills/nature-image2ppt/scripts/inspect_region_decomposition.py
@@ -373,7 +373,7 @@ def main() -> int:
         "errors": errors,
     }
     write_json(out, report)
-    print(json.dumps(report, ensure_ascii=False, indent=2))
+    print(json.dumps(report, ensure_ascii=True, indent=2))
     return 0 if report["passed"] else 2
