+++ skills/nature-image2ppt/scripts/run_final_image2ppt_qa.py
@@ -220,7 +220,7 @@ def main() -> int:
         "errors": errors,
     }
     write_json(paths["report"], report)
-    print(json.dumps(report, ensure_ascii=False, indent=2))
+    print(json.dumps(report, ensure_ascii=True, indent=2))
     return 0 if passed else 2
